remove static position prop. create Navbar folder and move files to the folder.
