--- src/components/Form.jsx
@@ -1,82 +1,0 @@
-import * as React from "react";
-import Box from "@mui/material/Box";
-import IconButton from "@mui/material/IconButton";
-import OutlinedInput from "@mui/material/OutlinedInput";
-import InputLabel from "@mui/material/InputLabel";
-import InputAdornment from "@mui/material/InputAdornment";
-import FormHelperText from "@mui/material/FormHelperText";
-import FormControl from "@mui/material/FormControl";
-import TextField from "@mui/material/TextField";
-import Visibility from "@mui/icons-material/Visibility";
-import { inputBaseClasses } from "@mui/material/InputBase";
-
-function Form() {
-  return (
-    <Box sx={{ display: "flex", flexWrap: "wrap" }}>
-      <div>
-        <TextField
-          id="outlined-suffix-shrink"
-          label="Outlined"
-          variant="outlined"
-          slotProps={{
-            input: {
-              endAdornment: (
-                <InputAdornment
-                  position="end"
-                  sx={{
-                    opacity: 0,
-                    pointerEvents: "none",
-                    [`[data-shrink=true] ~ .${inputBaseClasses.root} > &`]: {
-                      opacity: 1,
-                    },
-                  }}
-                >
-                  lbs
-                </InputAdornment>
-              ),
-            },
-          }}
-        />
-        <FormControl sx={{ m: 1, width: "25ch" }} variant="outlined">
-          <OutlinedInput
-            id="outlined-adornment-weight"
-            endAdornment={<InputAdornment position="end">kg</InputAdornment>}
-            aria-describedby="outlined-weight-helper-text"
-            inputProps={{
-              "aria-label": "weight",
-            }}
-          />
-          <FormHelperText id="outlined-weight-helper-text">
-            Weight
-          </FormHelperText>
-        </FormControl>
-        <FormControl sx={{ m: 1, width: "25ch" }} variant="outlined">
-          <InputLabel htmlFor="outlined-adornment-password">
-            Password
-          </InputLabel>
-          <OutlinedInput
-            id="outlined-adornment-password"
-            endAdornment={
-              <InputAdornment position="end">
-                <IconButton edge="end">
-                  <Visibility />
-                </IconButton>
-              </InputAdornment>
-            }
-            label="Password"
-          />
-        </FormControl>
-        <FormControl fullWidth sx={{ m: 1 }}>
-          <InputLabel htmlFor="outlined-adornment-amount">Amount</InputLabel>
-          <OutlinedInput
-            id="outlined-adornment-amount"
-            startAdornment={<InputAdornment position="start">$</InputAdornment>}
-            label="Amount"
-          />
-        </FormControl>
-      </div>
-    </Box>
-  );
-}
-
-export default Form;
