+++ src/components/Form/Form.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Form() {
+  return <div></div>;
+}
+
+export default Form;
