+++ src/components/Navbar/NavbarStyle.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+import { styled } from "@mui/material/styles";
+
+function NavbarStyle() {
+  return <div></div>;
+}
+
+export default NavbarStyle;
